requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
google-cloud-speech>=2.21.0
google-generativeai>=0.3.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
pydub>=0.25.1
aiofiles>=23.2.1
